# Further optimizations to tree