--- a/Makefile
+++ b/Makefile
@@ -65,5 +65,5 @@ test-watch:
 
 bench: javascript
 	@ node --expose-gc --trace-deopt benchmarks/tree-performance
-	@ node --expose-gc --trace-deopt benchmarks/dispatch-performance
-	@ node --expose-gc --trace-deopt benchmarks/push-performance
+	@ node --trace-deopt benchmarks/dispatch-performance
+	@ node --trace-deopt benchmarks/push-performance